Desarollado crud de Listas

+++ app/src/main/java/com/arribas/myshoppinglist/ui/view/shoplist/shoplistList/ShoplistListViewModel.kt
@@ -1,15 +1,27 @@
 package com.arribas.myshoppinglist.ui.view.shoplist.shoplistList
 
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.arribas.myshoppinglist.data.model.ArticleCategory
 import com.arribas.myshoppinglist.data.model.QArticle
 import com.arribas.myshoppinglist.data.model.Shoplist
 import com.arribas.myshoppinglist.data.repository.ShoplistRepository
 import com.arribas.myshoppinglist.data.utils.DIALOG_UI_TAG
+import com.arribas.myshoppinglist.data.utils.DialogUiState
+import com.arribas.myshoppinglist.ui.view.shoplist.ShoplistUiState
+import com.arribas.myshoppinglist.ui.view.shoplist.toItem
 import com.arribas.myshoppinglist.ui.view.shoplistList.SearchUiState
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asSharedFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
+import java.lang.Exception
 
 class ShoplistListViewModel(
     private val shoplistRepository: ShoplistRepository
@@ -21,27 +33,47 @@ class ShoplistListViewModel(
     private val _listUiState = MutableStateFlow(ShoplisListUiState())
     val listUiState: StateFlow<ShoplisListUiState> = _listUiState
 
+    /**AlertDialog functions****************************************/
+    private val _dialogState = MutableStateFlow(DialogUiState())
+    val dialogState: StateFlow<DialogUiState> = _dialogState.asStateFlow()
+
+    lateinit var shoplist: Shoplist
+
+    var shoplistUiState by mutableStateOf(ShoplistUiState())
+        private set
+
+    private val _toastMessage = MutableSharedFlow<String>()
+    val toastMessage = _toastMessage.asSharedFlow()
+
+    fun sendMessage(message: String) {
+        viewModelScope.launch {
+            _toastMessage.emit(message)
+        }
+    }
+
     init{
         getData()
     }
 
     fun search(_name: String){
-        //println("article $_name")
         _searchUiState.value = _searchUiState.value.copy(name = _name)
         getData()
     }
 
     fun clearName(){
-        println("article clearName")
         _searchUiState.value = _searchUiState.value.copy(name = "")
         getData()
     }
 
     fun onDialogDelete(shoplist: Shoplist){
-        /*onOpenDialogClicked(
-            qArticle = article,
+        onOpenDialogClicked(
+            shoplist = shoplist,
             tag = DIALOG_UI_TAG.TAG_DELETE
-        )*/
+        )
+    }
+
+    fun onUpdateItem(shoplist: Shoplist){
+        this.shoplistUiState.copy(id = shoplist.id, name = shoplist.name, type = shoplist.type)
     }
 
     fun updateItem(shoplist: Shoplist) {
@@ -73,9 +105,9 @@ class ShoplistListViewModel(
         }*/
     }
 
-    private fun onOpenDialogClicked(qArticle: QArticle? = null, tag: DIALOG_UI_TAG) {
-        /*if(qArticle !== null) {
-            article = qArticle
+    private fun onOpenDialogClicked(shoplist: Shoplist? = null, tag: DIALOG_UI_TAG) {
+        if(shoplist !== null) {
+            this.shoplist = shoplist
         }
 
         val _dialog: DialogUiState
@@ -85,7 +117,7 @@ class ShoplistListViewModel(
                 _dialog = DialogUiState(
                     tag = tag,
                     title = "¿Seguro que quieres continuar",
-                    body = "Si continuas se eliminará el artículo de la lista",
+                    body = "Si continuas se eliminará el item de la lista",
                     isShow = true,
                     isShowBtDismiss = true
                 )
@@ -93,20 +125,40 @@ class ShoplistListViewModel(
                 _dialog = DialogUiState()
         }
 
-        _dialogState.value = _dialog*/
+        _dialogState.value = _dialog
+    }
+
+    private fun deleteItem() {
+        if(shoplist == null){
+            return
+        }
+
+        viewModelScope.launch(Dispatchers.IO) {
+            //val articleCategorys = articleCategoryRepository.getByArticle(article.id)
+            //TODO:: Delete articles to list
+
+            try {
+                shoplistRepository.deleteItem(shoplist!!)
+
+                sendMessage("Item borrado correctamente")
+
+            }catch (e: Exception){
+                sendMessage("Se ha producido un error al borrar el item")
+            }
+        }
     }
 
     fun onDialogConfirm() {
-        /*when(_dialogState.value.tag) {
+        when(_dialogState.value.tag) {
             DIALOG_UI_TAG.TAG_DELETE  -> deleteItem()
             else -> {}
         }
 
-        _dialogState.value = DialogUiState(isShow = false)*/
+        _dialogState.value = DialogUiState(isShow = false)
     }
 
     fun onDialogDismiss() {
-        //_dialogState.value = DialogUiState(isShow = false)
+        _dialogState.value = DialogUiState(isShow = false)
     }
 
     private fun getData(){
